Add Logging. Start robust mode. Improve function arg names Added logging modules correctly to log to both a file and console. Could use fine tuning for the levels.
Started implementation of robust mode. Robust mode is an optional argument but defaults to True. To speed things up, turn it off.
Robust mode will check any setters with a getter. If it fails, it can auto-retry.
The same could be done for reports where auto-retry could be added.
The code works, but it could be a lot cleaner. It's difficult because each setter
and getter has different expected behavior.

Functions now will start becoming all keyword args instead of positional args.