Add http server functionality
This has basic capabilites for the web server to serve a page.
It allows filling out a wifi credential form, reads the HTTP Get,
and parses the results.

The wifi mode is currently station just so I can test without having to rejoin the wifi network.
Next, it will run the web server on softAP, connect to the network using the configured credentials,
save the credentials in flash if sucessful connect, and then set up MQTT on the station side.