Attempted fix #17
This issue, I think, was caused by WiFi. It manifested as device exceptions randomly when running. It was worse on ESP32D chip. It also corrupted the ESP32 when switching from AP to STA mode. The main fix was adding in some delays in the WiFi setup. Needs test on ESP32D