Support redirect on WiFi Config Finish
The redirect time depends on which URI is requested. The web ui serves a form that links to the 3 second redirect.
The 'instant' uri skips any redirect delays.

Solves #27

Note: Sometimes there are socket erros when sending the new web page. It's probably a timing issue with the transition to STA